refactor(commands): Implement modular command system and utilities
Refactors core components by introducing structured command mapping, process execution wrappers, and generalized utility functions. This change drastically improves code maintainability and extensibility for custom interaction commands.

Key changes include:
*   **Command System:** Added `include/commands.h` and associated logic to create a central map (`MappedCommands`) for registering and dispatching user-defined Discord interactions by name.
*   **Process Management:** Implemented dedicated functions in `src/process.cpp` and `include/process.h` to handle external process execution (`exec`) and child process creation (`createChild`), abstracting OS calls away from the main class logic.
*   **Utilities:** Moved general helper functions (e.g., character set generation, time retrieval, hashing) into dedicated modules in `src/utils.cpp` and `include/utils.h`.
*   **Architecture Update:** Modified `RABIDS` methods to utilize these new modular components, resulting in cleaner separation of concerns within the codebase.

This refactoring establishes a robust internal API boundary for user interaction logic and system utilities.
