+++ include/process.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <string>
+#include <vector>
+
+std::string exec(const char* cmd);
+int createChild(std::vector<std::string> arguments);
